Cancel a running countdown timer before starting it anew
Starting a new timer without clearing the previous one is not well-defined
as to behavior.  Will the old interval go away?  When does the interval
object become unreachable?  When does it stop calling its target function?

Setting this.interval to zero is not necessary since it will be set to
the new interval number before this method exits.

Signed-off-by: Sean Robinson <sean.robinson@scottsdalecc.edu>