Add radius getter/setter
Adjusting the radius on-the-fly will allow a Responsive UI.

This is a first draft of the new radius property, copied and reworked
from the monkey-patch version in the plumsul project.  Possible future
improvements include: re-factor _init to aid code re-use, simplify radius
method using re-factored _init code, and clean-up _getCanvas for
readability.

Signed-off-by: Sean Robinson <sean.robinson@scottsdalecc.edu>