Refactor _drawCountdownLabel to drawCountdownLabel
This is a major overhaul.  The input changed from secondsElapsed to
secondsLeft as the label shows remaining seconds.  Why internally
calculate that?  Let the caller tell us what to draw.

Use o as a shortcut for this.options.  This may be too short, but I'll
try this pattern for a while.

If this.options.label has a value, draw it.  There is no need to store
that status in another variable (i.e. drawLabel).

Only test for infinity if the remaining time is not one second.

Always start drawing at the center of the canvas.  The origin can be
relatively shifted, if needed.

Use template strings to build the font specifications in a more readable
way.  Not a major change, just more manageable into the future.

Signed-off-by: Sean Robinson <sean.robinson@scottsdalecc.edu>