Only call onTime callbacks once per second
In smooth mode, the draw method is called up to 60 times per second.  The
callbacks were also called many times per second.  As the intent of the
onTime callbacks is at most once per second, only call the callback if
this is the first call for this time slot.

While this rate limiting is active in course mode, it should have no
effect as it will only be checked once per second.

Signed-off-by: Sean Robinson <sean.robinson@scottsdalecc.edu>