Update README.md
The README now includes the major changes made in the previous two years.
Yes, the jQuery version mentioned in Basic Usage is an older version, but
was the latest version at the point of last active work on this project.
Later jQuery versions probably work just as well, but are untested.

Signed-off-by: Sean Robinson <sean.robinson@scottsdalecc.edu>